Fixed bound printing incorrectly for DFS* (with and without child ordering.)

+++ search/dfsstar-co.hpp
@@ -133,7 +133,7 @@ private:
 	// row outputs an incumbent solution row.
 	void row(unsigned long n, Cost cost) {
 		dfrow(stdout, "incumbent", "uuuggg", n, this->res.expd,
-			  this->res.gend, bound, (float)cost,
+			  this->res.gend, (float)bound, (float)cost,
 			walltime() - this->res.wallstart);
 	}
